Backend/python-server/requirements.txt
psycopg2==2.9.9
python-dotenv==1.0.1
rdflib==7.0.0
sentence_transformers==2.7.0